+++ evalml/__init__.py
@@ -21,7 +21,7 @@ import evalml.utils
 import evalml.guardrails
 
 from evalml.pipelines import list_model_types, save_pipeline, load_pipeline
-from evalml.models import AutoClassifier, AutoRegressor
+from evalml.models import AutoClassificationSearch, AutoRegressor
 
 warnings.filterwarnings("ignore", category=FutureWarning)
 warnings.filterwarnings("ignore", category=DeprecationWarning)