+++ evalml/models/__init__.py
@@ -4,5 +4,5 @@ Copyright (c) 2019 Feature Labs, Inc.
 The usage of this software is governed by the Feature Labs End User License Agreement available at https://www.featurelabs.com/eula/. If you do not agree to the terms set out in this agreement, do not use the software, and immediately contact Feature Labs or your supplier.
 """
 # flake8:noqas
-from .auto_classifier import AutoClassifier
+from .auto_classifier import AutoClassificationSearch
 from .auto_regressor import AutoRegressor