+++ evalml/tests/objective_tests/test_lead_scoring.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pandas as pd
 
-from evalml import AutoClassifier
+from evalml import AutoClassificationSearch
 from evalml.objectives import LeadScoring
 
 
@@ -11,7 +11,7 @@ def test_lead_scoring_objective(X_y):
     objective = LeadScoring(true_positives=1,
                             false_positives=-1)
 
-    clf = AutoClassifier(objective=objective, max_pipelines=1, random_state=0)
+    clf = AutoClassificationSearch(objective=objective, max_pipelines=1, random_state=0)
     clf.fit(X, y, raise_errors=True)
     pipeline = clf.best_pipeline
     pipeline.predict(X)